Add support for resource translation from on-disc to in-memory format. This allows loading and saving of resources without relying on the compiler's laying out structures in the exact way the old DOS compiler did. It's essential for 64-bit support. Map resources (type RTYPE_APP) use a table of data formats for each possible map index. Loading and saving of map resources now all go through this table, simplifying the load and save code. Table for both version 11 and 12 maps included. Other resources use a table of known RTYPEs and their formats.

+++ src/Libraries/2D/Source/GR/grs.h
@@ -179,17 +179,18 @@ typedef union {
    } i;
 } grs_clip;
 
+// Font.
 typedef struct {
-	ushort id;
-	char dummy1[34];
-	short min;
-	short max;
-	char dummy2[32];
-	long cotptr;
-	long buf;
-	short w;
-	short h;
-	short off_tab[1];
+    ushort id;
+    char dummy1[34];
+    short min;
+    short max;
+    char dummy2[32];
+    int32_t cotptr;
+    int32_t buf;
+    short w;
+    short h;
+    short off_tab[1];
 } grs_font;
 
 /* structure for drawing context.  the context contains data about which
